Localhost: correctly identify error line for heredoc/nowdoc strings
The sniff was incorrectly identifying the error line for URLs inside heredoc/nowdoc. Considering the nowdoc string added as a test in this commit, the sniff was identifying the error lines as 41 and 42 instead of the correct 40 and 41.

This was happening because of an error in the logic of the `find_token_in_multiline_string()` method. As far as I could test, this method is actually not necessary as just passing `$stackPtr` to `addError()` should be enough for it to correctly identify the error line.

PHPCS tokenizes multi-line strings using one token per line. So `$stackPtr` will correspond to the correct error line (the line that contains the localhost URL).

In a sniff like `Offloading` it makes sense to use this method as it combines the content of multi-line strings using `TextStrings::getCompleteTextString()`, but that is not the case for `Localhost`.

Removing `find_token_in_multiline_string()` also meant that it was not necessary anymore to pass the `PREG_OFFSET_CAPTURE` flag to `preg_match_all()` as the information about the offset is not used anymore.