Normalize The Unlicense to SPDX identifier
According to https://spdx.org/licenses/
the full name is "The Unlicense" and the SPDX
identifier is "Unlicense".

Update get_normalized_license to transform
both "The Unlicense" and "Unlicense",
case-insensitive, to "Unlicense".

This allows "The Unlicense" to be recognized in
a plugin's readme.txt. It also fixes a case where
if a plugin file said "The Unlicense" but the
readme.txt said "Unlicense", Plugin Check would
report that there was a license mismatch.